requirements.txt
Django~=4.1.5
django-filter~=22.1
djangorestframework~=3.14.0
django-cors-headers~=3.13.0
django-crontab~=0.7.1
django-object-actions~=4.1.0
drf-nested-routers~=0.93.4
psycopg2-binary~=2.9.5
